% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clitable-package.R
\docType{package}
\name{clitable-package}
\alias{clitable}
\alias{clitable-package}
\title{clitable: Render tables in text format in the terminal.}
\description{
Render tables in text format in the terminal using ANSI strings thanks to the cli and crayon packages.
}
\section{Features}{

TODO
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kforner/clitable}
  \item Report bugs at \url{https://github.com/kforner/clitable/issues}
}

}
\author{
\strong{Maintainer}: Karl Forner \email{karl.forner@gmail.com} [copyright holder]

}
