% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clitable-package.R
\docType{package}
\name{clitable-package}
\alias{clitable}
\alias{clitable-package}
\title{clitable: Render Tables in Text for the Terminal}
\description{
Render tables in text format in the terminal using ANSI strings thanks to the 'cli' and 'crayon' packages.
}
\section{Features}{

\itemize{
\item can display any ansi string (without end of lines) content
\item multiple table border styles: single, double, single-double, double-single, classic
\item can display heatmaps
\item can highlight rows
\item can display NAs with custom style
\item few dependencies: only \code{crayon} and \code{cli}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kforner/clitable}
  \item Report bugs at \url{https://github.com/kforner/clitable/issues}
}

}
\author{
\strong{Maintainer}: Karl Forner \email{karl.forner@gmail.com} [copyright holder]

}
